refactor(kit): rewrite simple entity forms (realm/role/scope/permission) as SFCs
Convert four single-entity forms from hyperscript builder VNodes
to native <template> SFCs. Same exports / props / emits; the only
externally visible difference is the loss of the buildFormGroup /
buildFormInput / buildFormTextarea / buildFormSubmit shim chain.

Each form:
- Reads from useVuelidate and exposes the validation accessor as
  `vuelidate` on the setup return.
- Wraps each field in <IVuelidate><VCFormGroup>...</VCFormGroup>
  matching the AClientForm.vue pattern.
- Uses v-model on VCFormInput / VCFormTextarea / VCFormSelect for
  reactive binding (no more shim closure with stale onChange).
- Submit via <AFormSubmit @submit="submit"/>.
- Generate-name button and realm-picker fallback rendered with
  proper template conditionals.

Forms migrated: ARealmForm, ARoleForm, AScopeForm, APermissionForm.
