+++ packages/client-web-kit/src/components/entities/permission/index.ts
@@ -7,5 +7,5 @@
 
 export { APermission } from './APermission';
 export { APermissionCheck } from './APermissionCheck';
-export { APermissionForm } from './APermissionForm';
+export { APermissionForm } from './APermissionForm.vue';
 export { APermissions } from './APermissions';
