+++ packages/client-web-kit/src/components/entities/realm/index.ts
@@ -6,6 +6,6 @@
  */
 
 export { ARealm } from './ARealm';
-export { ARealmForm } from './ARealmForm';
+export { ARealmForm } from './ARealmForm.vue';
 export { ARealmPicker } from './ARealmPicker';
 export { ARealms } from './ARealms';
